\name{wed-package}
\alias{wed-package}
\alias{wed-package}
\docType{package}
\title{
WED 
}
\description{
Estimate weighted euclidean distance spatial model, 
allowing for separability 
}
\details{
\tabular{ll}{
Package: \tab wed\cr
Type: \tab Package\cr
Version: \tab 0\cr
Date: \tab 2014-04-09\cr
}
Most important function wed
}
\author{
Lukas Stoetzer, University of Mannheim \cr
Steffen Zittlau, University of Mannheim
}
\references{
Stoetzer and Zittlau: Estimating Nonseparable preferences in Mass Elections. 
}

